+++ src/data/cardChartData.js
@@ -1,0 +1,58 @@
+import {getStyle} from "@coreui/coreui/dist/js/coreui-utilities";
+const brandPrimary = getStyle('--primary');
+const brandInfo = getStyle('--info');
+
+const cardChartData1 = {
+  labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
+  datasets: [
+    {
+      label: 'My First dataset',
+      backgroundColor: brandPrimary,
+      borderColor: 'rgba(255,255,255,.55)',
+      data: [65, 59, 84, 84, 51, 55, 40],
+    },
+  ],
+};
+
+const cardChartData2 = {
+  labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
+  datasets: [
+    {
+      label: 'My First dataset',
+      backgroundColor: brandInfo,
+      borderColor: 'rgba(255,255,255,.55)',
+      data: [1, 18, 9, 17, 34, 22, 11],
+    },
+  ],
+};
+
+const cardChartData3 = {
+  labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
+  datasets: [
+    {
+      label: 'My First dataset',
+      backgroundColor: 'rgba(255,255,255,.2)',
+      borderColor: 'rgba(255,255,255,.55)',
+      data: [78, 81, 80, 45, 34, 12, 40],
+    },
+  ],
+};
+
+const cardChartData4 = {
+  labels: ['', '', '', '', '', '', '', '', '', '', '', '', '', '', '', ''],
+  datasets: [
+    {
+      label: 'My First dataset',
+      backgroundColor: 'rgba(255,255,255,.3)',
+      borderColor: 'transparent',
+      data: [78, 81, 80, 45, 34, 12, 40, 75, 34, 89, 32, 68, 54, 72, 18, 98],
+    },
+  ],
+};
+
+export {
+  cardChartData1,
+  cardChartData2,
+  cardChartData3,
+  cardChartData4
+}
