+++ src/data/mainChart.js
@@ -1,0 +1,53 @@
+import { getStyle, hexToRgba } from "@coreui/coreui/dist/js/coreui-utilities";
+const brandSuccess = getStyle('--success');
+const brandInfo = getStyle('--info');
+const brandDanger = getStyle('--danger');
+
+//Random Numbers
+const random = (min, max) => {
+  return Math.floor(Math.random() * (max - min + 1) + min);
+};
+
+let elements = 27;
+let data1 = [];
+let data2 = [];
+let data3 = [];
+
+for (let i = 0; i <= elements; i++) {
+  data1.push(random(50, 200));
+  data2.push(random(80, 100));
+  data3.push(65);
+}
+
+const mainChart = {
+  labels: ['Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa', 'Su', 'Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa', 'Su', 'Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa', 'Su', 'Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa', 'Su'],
+  datasets: [
+    {
+      label: 'My First dataset',
+      backgroundColor: hexToRgba(brandInfo, 10),
+      borderColor: brandInfo,
+      pointHoverBackgroundColor: '#fff',
+      borderWidth: 2,
+      data: data1,
+    },
+    {
+      label: 'My Second dataset',
+      backgroundColor: 'transparent',
+      borderColor: brandSuccess,
+      pointHoverBackgroundColor: '#fff',
+      borderWidth: 2,
+      data: data2,
+    },
+    {
+      label: 'My Third dataset',
+      backgroundColor: 'transparent',
+      borderColor: brandDanger,
+      pointHoverBackgroundColor: '#fff',
+      borderWidth: 1,
+      borderDash: [8, 5],
+      data: data3,
+    },
+  ],
+};
+
+export default mainChart;
