+++ src/data/sparkLineChartData.js
@@ -1,0 +1,27 @@
+export default [
+  {
+    data: [35, 23, 56, 22, 97, 23, 64],
+    label: 'New Clients',
+  },
+  {
+    data: [65, 59, 84, 84, 51, 55, 40],
+    label: 'Recurring Clients',
+  },
+  {
+    data: [35, 23, 56, 22, 97, 23, 64],
+    label: 'Pageviews',
+  },
+  {
+    data: [65, 59, 84, 84, 51, 55, 40],
+    label: 'Organic',
+  },
+  {
+    data: [78, 81, 80, 45, 34, 12, 40],
+    label: 'CTR',
+  },
+  {
+    data: [1, 13, 9, 17, 34, 41, 38],
+    label: 'Bounce Rate',
+  },
+];
+// sparkline charts
