+++ src/data/sparklineChartOpts.js
@@ -1,0 +1,36 @@
+import { CustomTooltips } from "@coreui/coreui-plugin-chartjs-custom-tooltips";
+
+const sparklineChartOpts = {
+  tooltips: {
+    enabled: false,
+    custom: CustomTooltips
+  },
+  responsive: true,
+  maintainAspectRatio: true,
+  scales: {
+    xAxes: [
+      {
+        display: false,
+      }],
+    yAxes: [
+      {
+        display: false,
+      }],
+  },
+  elements: {
+    line: {
+      borderWidth: 2,
+    },
+    point: {
+      radius: 0,
+      hitRadius: 10,
+      hoverRadius: 4,
+      hoverBorderWidth: 3,
+    },
+  },
+  legend: {
+    display: false,
+  },
+};
+
+export default sparklineChartOpts;
